LLaMa3-8B/requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bitsandbytes==0.43.3
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
datasets==2.20.0
dill==0.3.8
docstring_parser==0.16
fastapi==0.112.1
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
frozenlist==1.4.1
fsspec==2024.5.0
grpclib==0.4.7
h2==4.1.0
hf_transfer==0.1.8
hpack==4.0.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
mdurl==0.1.2
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
modal==0.64.42
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict==6.0.5
multiprocess==0.70.16
networkx @ file:///croot/networkx_1720002482208/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow @ file:///croot/pillow_1721059439630/work
protobuf==3.20.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
rich==13.7.1
safetensors==0.4.4
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
shtab==1.7.1
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy @ file:///croot/sympy_1701397643339/work
synchronicity==0.7.3
tensorboardX==2.6.2.2
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
trl==0.9.6
typer==0.12.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.8.6
tzdata==2024.1
unsloth @ git+https://github.com/unslothai/unsloth.git@a64b8f648ad067f9745253161e73a0367bf0ca5a
urllib3 @ file:///croot/urllib3_1718912636303/work
uv==0.2.35
watchfiles==0.23.0
wheel==0.43.0
xformers==0.0.27.post2
xxhash==3.4.1
yarl==1.9.4
